do not reinstall the npm world every time

+++ sideloader/postinstall.sh
@@ -2,9 +2,17 @@ pip="${VENV}/bin/pip"
 
 cd "${INSTALLDIR}/${REPO}/"
 
-npm install -g coffee-script
-npm install -g bower
-bower install lessc
+if [ ! -f `which coffee` ]; then
+    npm install -g coffee-script
+fi
+
+if [ ! -f `which bower` ]; then
+    npm install -g bower
+fi
+
+if [ ! -f `which lessc` ]; then
+    bower --allow-root install lessc
+fi
 
 $pip install -r pip-freeze.txt
-python manage.py syncdb
+PYTHONPATH=`pwd` python manage.py syncdb
